Update Koios Client API routing.
