+++ backend/app/api/main.py
@@ -1,6 +1,7 @@
+from backend.app.KoiosClient import ai_router
 from fastapi import APIRouter
 
-from app.api.routes import ai, checkins, engine89, items, login, orientations, private, questionnaires, users, utils
+from app.api.routes import checkins, engine89, items, login, orientations, private, questionnaires, users, utils
 from app.core.config import settings
 from app.goal_scaffold.router import goal_scaffold_router
 
@@ -12,7 +13,7 @@ api_router.include_router(items.router)
 api_router.include_router(orientations.router)
 api_router.include_router(questionnaires.router, prefix="/questionnaires", tags=["questionnaires"])
 api_router.include_router(goal_scaffold_router, prefix="/scaffold", tags=["goal-scaffold"])
-api_router.include_router(ai.router)
+api_router.include_router(ai_router)
 api_router.include_router(engine89.router)
 api_router.include_router(checkins.router)
